backend-core split wip
